[Fix] Update operators.py
Fixed a possible error on line 2021: "TypeError: 'numpy.float64' object cannot be interpreted as an integer".

+++ ppdet/data/transform/operators.py
@@ -2032,8 +2032,8 @@ class Pad(BaseOperator):
                 im_h < h and im_w < w
             ), '(h, w) of target size should be greater than (im_h, im_w)'
         else:
-            h = np.ceil(im_h / self.size_divisor) * self.size_divisor
-            w = np.ceil(im_w / self.size_divisor) * self.size_divisor
+            h = int(np.ceil(im_h / self.size_divisor) * self.size_divisor)
+            w = int(np.ceil(im_w / self.size_divisor) * self.size_divisor)
 
         if h == im_h and w == im_w:
             return sample
